apps/api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
Wand==0.6.13
pydantic==2.8.2
pydantic-settings==2.3.4
sqlalchemy==2.0.34
psycopg[binary]==3.2.1
alembic==1.13.1
python-multipart==0.0.9
boto3==1.34.162
pypdf==4.3.1
pillow==10.4.0
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
structlog==24.1.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.27.0
pdf2image==1.17.0
google-generativeai==0.8.3
duckduckgo-search==4.1.1
